Update txt2tags, modify markdown conversion (#30)
This PR updates text2tags version 3.9.
This is to support converting t2t to markdown.
See also nvaccess/nvda#15945

It also applies a small patch updating the rules for markdown.

There is no universal standard for custom anchors in markdown.
As such text2tags doesn't have default rules for this. To retain our custom anchors, I added rules for a common markdown extended syntax. This is supported by the python markdown legacy attributes extension.
Similarly we use a custom syntax for autogenerating the Table of contents.

We also need to fix the markdown conversion so that the page title (first line of t2t file) has heading markup.

Numbered titles are also stripped in favour of numberless titles.